fixed layout in auth
